Moved NK Auth components to their own directory
