+++ src/main.tsx
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom/client";
 import App from "./App.tsx";
 import "./index.css";
 import HttpsRedirect from "react-https-redirect";
-import NkAuthProvider from "./NkAuthProvider.tsx";
+import NkAuthProvider from "./nk-auth/NkAuthProvider.tsx";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
